+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added option `config.WithoutDNSResolver` for exclude internal dns-resolver and use direct FQDN for dialing
+
 ## v3.15.1
 * Removed all conditions for trace retry errors
 * Fixed background color of warn messages